Move DataTable and StringColumn integrity checks into file integrity_checks.cc
